syakuis/ainframe#11
MenuContext 에서 하나의 메뉴뿐아니라 전체를 읽을 수 있게 수정
전체 데이터를 한번에 읽을 필요성이 없음. 현재는 메뉴명과 메뉴번호만 리스트업해서 선택할 수 있는 데이터가 필요함. 이부분으로 개선함.

추가로 파이어버그 관련 사항 처리

+++ ainframe-core/src/main/java/org/ainframe/core/util/PageNavigator.java
@@ -78,7 +78,7 @@ public class PageNavigator {
         navigator.startPage = (int) Math.ceil(((page - 1) / pageLink) * pageLink + 1);
         navigator.endPage = navigator.startPage + pageLink - 1;
 
-        navigator.virtualIdx = navigator.totalCount - (pageRow * (page - 1));
+        navigator.virtualIdx = navigator.totalCount - (long) (pageRow * (page - 1));
         // oracle start rowNum = 0
         //virtualIdx = virtualIdx + 1;
